Premailer v1.0.0 Initial Commit
With maintainability in mind, the default export Premailer class is backed with 'fs' based data validations and an Options class to validate user-defined premailerOptions. All operational errors are accounted for in this initial commit (version 1.0.0).

The basis of the premailer-process has been created. However, this will need a revision to include other key features that a premailer needs, such as:
- implementing premailerOptions.makeGhostTables to add compatibility for the 'max-width' css property.
- Adding conditional style and xml tags for detected mso versions.
- Likewise, adding styles tags that are targeted to certain screen sizes, to replace media queries.
- Swapping specific CSS properties for their HTML attribute alternatives (e.g. 'padding' for 'cellpadding' and 'border-spacing' for 'cellspacing').

I will need to research premailer processes further, and patch known bugs in my code. I will also refacter my premailer-process files in the next revision, to make the code more efficient, readable and maintainable.